move the color methods from model to the view.
